Begin rewrite
