+++ Pipfile
@@ -1,0 +1,14 @@
+[[source]]
+name = "pypi"
+url = "https://pypi.org/simple"
+verify_ssl = true
+
+[dev-packages]
+
+[packages]
+python-twitter = "*"
+python-dotenv = "*"
+instagram-private-api = {git = "https://git@github.com/ping/instagram_private_api.git",ref = "1.6.0"}
+
+[requires]
+python_version = "3.6"
