+++ config.py
@@ -1,0 +1,3 @@
+import os
+
+LAST_FILE_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'last')
