--- requirements.txt
@@ -1,11 +1,0 @@
-beautifulsoup4==4.6.0
-certifi==2018.4.16
-chardet==3.0.4
-future==0.16.0
-idna==2.6
-oauthlib==2.1.0
-python-dotenv==0.8.2
-python-twitter==3.4.1
-requests==2.18.4
-requests-oauthlib==0.8.0
-urllib3==1.22
